+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### V 1.5.0
 
-- Add Phone orientation trigger
+- Add phone orientation trigger
+- Add noise trigger
 - Lots of backend changes to support Watches
 - Rework permissions
 - Remove marketing notifications